Image Composer page system now takes account of ICGPUConstants, allowing different sharpness for fonts. allowing multiple fonts at same layer of various sizes at highest quality (before might get the wrong sharpness) General tidying, moving to the C++11 for range format for example

+++ src/libs/gui/swfruntime/builder.cpp
@@ -272,8 +272,8 @@ namespace Swf {
 		// morphs always use direct quadratic rendering rather than the subdiving renderer
 		bool useDirectQuadratic = false; // TODO player.UseGpuDirectQuadraticDrawer | _morphShape;
 
-		int fillLen = _fillStyles.size();
-		int lineLen = _lineFillStyles.size();
+		int fillLen = (int)_fillStyles.size();
+		int lineLen = (int)_lineFillStyles.size();
 		
 		if( !useDirectQuadratic ){
 			_fillPaths.resize( fillLen );
